+++ websites/management/commands/markdown_cleaning/shortcode_parser.py
