+++ websites/management/commands/markdown_cleaning/shortcode_parser_test.py
@@ -1,6 +1,6 @@
 import pytest
 
-from websites.management.commands.markdown_cleaning.shortcode_grammar import (
+from websites.management.commands.markdown_cleaning.shortcode_parser import (
     ShortcodeParam,
     ShortcodeParser,
     ShortcodeTag,
